Use Type[] instead of Class<?>[] for storing parameter types of constructor arguments
Lazy component resolution is harder using Class<?> as this does not store the information of the generic type of the Lazy class. This only allows us to know that one of the parameters in the constructor is Lazy, but we don't know which component to initialize as lazy. Using Type on the other hand keeps the information of the generic type which lets us determine the component that needs to be initialized lazy.