Fix interface classes in parameters not being resolved properly
Previously when trying to resolve a parameter of a constructor that was an interface, it would throw an error as we would try to create a definition for an interface and no constructors would be available. Now we register all the metadata for the scanned classes and when trying to resolve an interface parameter we check all the definitions existent for the given interface and resolve those first and then we can proceed with the injection into the constructor.